shopcart-service/pyproject.toml
[project]
name = "shopcart_service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.17.0",
    "fastapi[standard]>=0.118.0",
    "psycopg[binary]>=3.2.10",
    "pydantic-settings>=2.11.0",
    "sqlalchemy>=2.0.43",
    "uvicorn[standard]>=0.37.0",
]
